+++ sql/rls.sql
@@ -15,3 +15,9 @@ FROM
 WHERE
 	tablename = 'profiles';
 
+-- the policy below is needed to allow update calls, since the user not only needs to do a update but also a select
+CREATE POLICY "Enable update for authenticated users only with select" ON "public"."processes" AS permissive
+	FOR UPDATE TO authenticated
+		WITH CHECK (TRUE)
+		USING (TRUE);
+