+++ CHANGELOG.md
@@ -1,3 +1,55 @@
+# [1.1.0-rc.6](https://github.com/technologiestiftung/zms-supabase/compare/v1.1.0-rc.5...v1.1.0-rc.6) (2022-10-19)
+
+
+### Bug Fixes
+
+* Function parameter naming error in migration ([e6754df](https://github.com/technologiestiftung/zms-supabase/commit/e6754df64a3b142651bd7eae0bdc197102a0a12e))
+
+
+### Features
+
+* **relation:** Add service type relation table ([00dfd21](https://github.com/technologiestiftung/zms-supabase/commit/00dfd21eb399c7eb8bd99d17a81154947b894b04))
+
+# [1.1.0-rc.5](https://github.com/technologiestiftung/zms-supabase/compare/v1.1.0-rc.4...v1.1.0-rc.5) (2022-10-19)
+
+
+### Features
+
+* Drop column active from processes table ([e260107](https://github.com/technologiestiftung/zms-supabase/commit/e260107ac6a94c23c5a769fb2df10b8f2a4eb5db))
+
+# [1.1.0-rc.4](https://github.com/technologiestiftung/zms-supabase/compare/v1.1.0-rc.3...v1.1.0-rc.4) (2022-10-19)
+
+
+### Bug Fixes
+
+* RLS for update ([ea6b9d8](https://github.com/technologiestiftung/zms-supabase/commit/ea6b9d836485a4360b74a70ce3aabfbc097295a8))
+
+# [1.1.0-rc.3](https://github.com/technologiestiftung/zms-supabase/compare/v1.1.0-rc.2...v1.1.0-rc.3) (2022-10-19)
+
+
+### Bug Fixes
+
+* migration ([9782981](https://github.com/technologiestiftung/zms-supabase/commit/9782981d1a92c0b31e52f92fbf823b3ccb20577b))
+
+
+### Features
+
+* Add function to compute scores ([0e35500](https://github.com/technologiestiftung/zms-supabase/commit/0e355005b5c2fb49b708beccecfde98521b2618c))
+
+# [1.1.0-rc.2](https://github.com/technologiestiftung/zms-supabase/compare/v1.1.0-rc.1...v1.1.0-rc.2) (2022-10-18)
+
+
+### Bug Fixes
+
+* User deletion from auth schema ([8834e5c](https://github.com/technologiestiftung/zms-supabase/commit/8834e5c5241f653c557a513584ac7985f1333688))
+
+# [1.1.0-rc.1](https://github.com/technologiestiftung/zms-supabase/compare/v1.0.1...v1.1.0-rc.1) (2022-10-17)
+
+
+### Features
+
+* Add tabs for checkin and desk service ([9404893](https://github.com/technologiestiftung/zms-supabase/commit/94048939bf587bf3822f271a235e1ae654eb009b))
+
 ## [1.0.1](https://github.com/technologiestiftung/zms-supabase/compare/v1.0.0...v1.0.1) (2022-10-05)